requirements.txt
asgiref==3.3.1
Django==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
stories==4.0.0
uWSGI==2.0.19.1